+++ AwsConfig.psm1
@@ -136,13 +136,26 @@ function Invoke-AwsCli {
         [string[]]$Arguments
     )
 
-    $output = & aws @Arguments
+    $combined = & aws @Arguments 2>&1
     $exitCode = [int]$LASTEXITCODE
-    $joined = if ($null -eq $output) { '' } else { ($output -join [Environment]::NewLine) }
+
+    $stdoutLines = New-Object System.Collections.Generic.List[string]
+    $stderrLines = New-Object System.Collections.Generic.List[string]
+    foreach ($item in $combined) {
+        if ($item -is [System.Management.Automation.ErrorRecord]) {
+            $stderrLines.Add($item.ToString())
+        } else {
+            $stdoutLines.Add([string]$item)
+        }
+    }
+
+    $joined       = if ($stdoutLines.Count -eq 0) { '' } else { ($stdoutLines.ToArray() -join [Environment]::NewLine) }
+    $stderrJoined = if ($stderrLines.Count -eq 0)  { '' } else { ($stderrLines.ToArray()  -join [Environment]::NewLine) }
 
     return [PSCustomObject]@{
         ExitCode = $exitCode
         Output   = $joined
+        Stderr   = $stderrJoined
         Success  = ($exitCode -eq 0)
     }
 }